+++ freelens/package.json
@@ -141,7 +141,6 @@
     "typed-regex": "^0.0.8",
     "url-parse": "^1.5.10",
     "uuid": "^11.1.0",
-    "win-ca": "^3.5.1",
     "winston": "^3.17.0",
     "ws": "^8.18.3",
     "zod": "^3.25.67"
