+++ packages/core/package.json
@@ -179,7 +179,6 @@
     "typed-regex": "^0.0.8",
     "url-parse": "^1.5.10",
     "uuid": "^11.1.0",
-    "win-ca": "^3.5.1",
     "winston": "^3.17.0",
     "winston-transport": "^4.9.0",
     "winston-transport-browserconsole": "^1.0.5",
@@ -228,7 +227,6 @@
     "@types/uuid": "^10.0.0",
     "@types/webpack-env": "^1.18.8",
     "@types/webpack-node-externals": "^3.0.4",
-    "@types/win-ca": "^3.5.4",
     "@types/ws": "^8.18.1",
     "circular-dependency-plugin": "^5.2.2",
     "css-loader": "^6.11.0",
