+++ pnpm-workspace.yaml
@@ -11,7 +11,6 @@ onlyBuiltDependencies:
   - node-pty
   - playwright
   - sharp
-  - win-ca
 
 overrides:
   http-proxy: npm:http-proxy-node16@^1.0.0
